Fix end of month bugs
At the end of the month, certain bugs with the display of the timegraph
would pop up.

Simplified monthPercentage creation code.

Fix bug where green would not show on the last line. This was because
I was reseting or displaying red instead of checking if it was
originally green and needing to be displayed as such.

Fix bug where if at the end of the month the last char would be
duplicated. This is now fixed by checking if the percentage is 10, then
deleting the duplicate entry.

Signed-off-by: userWayneCampbell <wcampbell1995@gmail.com>